court-value/requirements.txt
fastapi==0.115.12
uvicorn==0.30.6
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
xgboost==2.1.1
shap==0.46.0
pydantic==2.9.2
pytest==8.3.3
nba_api==1.7.0
requests==2.32.3
python-dotenv==1.0.1
pyarrow==17.0.0
sqlalchemy==2.0.36
